Changed the name display functionality to display on clicking vs. mouseover
Changed the code such that rather than the name of the person in the graph displaying when the mouse is over their bar graph, it now appears once clicked. May consider finding a way to make the name diasappear when clicking off the graph. Removed console.log statement that was used to help debug.