Removed unessecary state variable and cleared console error
No longer using this.props.click as it was unessecary and the code can be reduced by just looking at the person's name. Only displays the name once there is a name (length >0 and now set to 0 rather than null). Fixed compiler error by removing the fetch for "api/people/".